go/types: don't set a Config.Context if none is provided
Users can re-use a type checking context by passing it via types.Config.
There is no need for us to expose the internal type checking context
when the config context is unset, and in fact doing so could lead to a
memory leak for users that re-use types.Config, expecting it to be small
and immutable.

Keep track of the Context on Checker instead, and zero it out at the end
of type checking and start of calls to Checker.Files.

For good measure, also zero out some other recently added Checker state
at the start of Checker.Files

Change-Id: Iff5b328a09cd0af76fcd4869f5f15352131b5986