os/exec: avoid NewFile on unknown FDs
exec_test.go's init function uses os.NewFile(fd) + f.Stat as a portable
mechanism to determine if an FD is in use.

Unfortunately, the current use is racy: if an unused FD becomes used
between NewFile and f.Close, then we will unintentionally close an FD we
do not use.

We cannot simply drop Close, as the finalizer will close the FD. We
could hold all of the os.Files in a global for the lifetime of the
process, but the need for such a hack is indicative of the larger
problem: we should not create an os.File for an FD that we do not own.

Most OSes provide a simple FS directory containing all open FDs for the
process. For these OSes, use a stat in that directory to check for FD
use.

The test using this data is already OS-independent, so it will lose a
bit of coverage from the OSes we don't list here, but that isn't
fundamental to the test.

Fixes #49533

Change-Id: I36e2bdb15f271ab01e55c18db6564271995a15af