cmd/link: add some better DWARF location expression tests
This patch rolls out a new set of test scaffold for doing regression
testing of DWARF variable locations. The scaffold consists of a
harness program 'dwdumploc' that can be run on a Go binary to dump out
the DWARF location expressions for a function's formal parameters, and
then a new set of tests that invoke the harness on specific canned
testcases.

DWARF location expressions can refer to machine registers, making the
output of the dump architecture-dependent. The dumper program
currently supports { Amd64, Arm64 } x { Linux, Macho, Windows }.

Updates #47354.

Change-Id: Ie7b90b58068a8df2137a614b21066254755c7549