cmd/internal/obj/arm64: add assembly syntax for the SVE
The SVE (Scalable Vector Extension) is a new instruction set,
and some SVE-specific argument types are newly introduced.
But their Go assembly syntax is not defined, so we first add
mapping rules for SVE before implementing them, so that if
users want to add the support in the assembler, there can be
a clear and consistent pattern.

We posted a discussion of SVE assembly syntax to the golang-dev
mailing list (https://groups.google.com/forum/#!topic/golang-dev/sYctj3HrXvc),
this patch uses the recommended format.

SVE related documents can be found on:
1. https://developer.arm.com/tools-and-software/server-and-hpc/compile/arm-instruction-emulator/resources/tutorials/sve
2. https://developer.arm.com/documentation/ddi0602/2021-09/SVE-Instructions?lang=en

Change-Id: I95649dcffc3309722430b6a1ba215aa414f909ac