net: check error from syscall.Getsockname in (*netFD).dial
The error from the getsockname system call was being ignored; I'm not
sure why, because 'man getsockname' lists a number of possible failure
modes, including ENOBUFS (for resource exhaustion) and EBADF (which we
could possibly see in the event of a bug or race condition elsewhere
in the program).

Now we explicitly check that error and fail the dial call if it is
non-nil. This probably doesn't fix #34611, but may help us diagnose it
more clearly.

While we're add it, also add more nil-checking logic in the test based
on the stack traces posted to
https://golang.org/issue/34611#issuecomment-975923748.

For #34611

Change-Id: Iba870b96787811e4b9959b74ef648afce9316602