log/syslog: create unix sockets in the test temp directory, not the system default
startServer invokes os.Remove on the temporary file for a unix socket
after creating it. If files are created in the global temp directory,
that can cause two tests to arrive at colliding names.

(Noticed while looking into the failure at
https://storage.googleapis.com/go-build-log/af2c83b1/solaris-amd64-oraclerel_3e01fda8.log,
but I would be surprised if this solves that failure.)

Change-Id: I21743f245e5c14645e03f09795013e058b984471