go/types: eliminate unnecessary complexity in initOrder
Our calculation of initOrder builds the dependency graph and then
contracts function edges approximately at random. While profiling, I
noticed that this latter step introduces a superlinear algorithm into
our type checking pass, which can dominate type checking for large
packages such as runtime.

To illustrate, suppose the following assumptions hold:
- Every function makes D calls at random to other functions in the
  package, for some fixed constant D.
- The number of functions is proportional to N, the size of the package.

Under these simplified assumptions, the cost of contracting an arbitrary
function F is P*D, where P is the expected number of functions calling
F. P has a poisson distribution with mean D.

Now consider the fact that when contracting a function F in position i,
we recursively pay the cost of copying F's predecessors and successors
for each node in the remaining uncontracted subgraph of functions
containing F. With our assumptions, the size of this subgraph is
proportional to (N-i), the number of remaining functions to contract.

Therefore, the total cost of contracting functions is proportional to

  P*D*Σᴺ(N-i)

which is proportional to N².

However, if we contract functions in ascending order of cost, we can
partition by the number of predecessors, and it can be seen that the
total cost of contracting functions is proportional to

  N*D*Σ(PMF(X))

where PMF is the probability mass function of P. I.e. cost is
proportional to N.

Assuming the above analysis is correct, it is still the case that the
initial assumptions are naive. Many large packages are more effectively
characterized as combinations of smaller packages. Nevertheless, it is
intuitively clear that contracting expensive edges last should be
cheaper.

Therefore, we add a new dependencyGraph2 algorithm which sorts functions
by increasing cost. For most packages this doesn't make a noticeable
difference, but for runtime the difference is large:

name                               old time/op  new time/op  delta
Check/runtime/funcbodies/noinfo-8   568ms ±25%    82ms ± 1%   -85.53%  (p=0.000 n=8+10)

name                               old lines/s  new lines/s  delta
Check/runtime/funcbodies/noinfo-8   93.1k ±56%  705.1k ± 1%  +657.63%  (p=0.000 n=10+10)

For now, I have kept the old algorithm so that it may be compared for
correctness when debug=true.

Updates #49856

Change-Id: Id2e70d67401af19205e1e0b9947baa16dd6506f0