cmd/compile/internal/syntax: fix parsing of type parameter lists
The parser cannot distinguish a type parameter list of the form

[P *T ] or
[P (T)]

from an array length specification P*T (product) or P(T)
(constant-valued function call) and thus interprets these
forms as the start of array types. This ambiguity must be
resolved explicitly by placing *T inside an interface, or
by leaving the parentheses away, respectively.

This CL adjusts the parser such that these forms are
interpreted as (the beginning) of type parameter lists
if the token after P*T or P(T) is a comma, or if T is
a type literal.

This CL also adjusts the printer to print a comma if
necessary to avoid this ambiguity.

Fixes #49482

Change-Id: I36328e2a7d9439c39ba0349837c445542549e84e