cmd/api: read list of ports from 'go tool dist list -json' for api check
'go tool dist test -run api' now uses the list of ports returned by 'go tool dist list -json'.

One file with features added for almost every port is added.
Most of them are for the syscall package.

The release used to name the files is from the release notes.
When cgo was not mentioned, the file of the cgo port is created next to the non-cgo file.

Ports removed for which no API was added (nacl) will not appear.
Exceptions file is supplemented with feature added before go1.1 and implemented later for every port.

Fixes #14892.

Change-Id: Id15fd5a5115820c7576de813f5660db306d5bd03