cmd/api: set default to first class ports.
Ports are listed using 'go tool dist list -json' which is exhaustive. Default, like used
on the builders, is now the list of first class ports. By default, without a specific context,
 a missing feature for a non-first class port is ignored.

Files for ports promoted to first class are added in the api folder. Otherwise,
their features would appear as added when using defaults.

When using default, features which are common but only found in port-specific files
 appear as added. A very small file is added as Go1.1 to avoid an irrelevant error.

Fixes #14892.

Change-Id: Id15fd5a5115820c7576de813f5660db306d5bd03